make selector private
